NonMCPMode/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
aiohttp>=3.8.1
aioconsole>=0.3.3
rich>=10.12.0
PyYAML>=6.0
torch>=2.0.0
transformers>=4.30.0
python-multipart>=0.0.5
pyyaml>=6.0.1
